cli/pyproject.toml
[tool.poetry]
name = "cli"
version = "0.1.0"
description = ""
authors = ["Carl Csaposs <carl.csaposs@canonical.com>"]
readme = "README.md"

[tool.poetry.scripts]
collect-charms = "cli.collect_charms:main"
collect-bases = "cli.collect_bases:main"
build = "cli.build:main"
release = "cli.release:main"
add-charm = "cli.add_charm_branch:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
requests = "^2.32.3"
uritemplate = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I", "UP"]
